+++ lib/expire_it.dart
@@ -2,3 +2,5 @@ library expire_it;
 
 export 'src/expire_builder.dart';
 export 'src/user_define_expire_controller.dart';
+export 'src/disabled_expire_controller.dart';
+export 'src/expire_it.dart';