+++ lib/src/expire_builder.dart
@@ -1,6 +1,7 @@
 import 'package:expire_it/src/internal/expire_base_controller.dart';
 import 'package:flutter/material.dart';
-import 'expire_state.dart';
+import 'internal/expire_it_provider.dart';
+import 'internal/expire_state.dart';
 
 /// A widget that dynamically builds its content based on the state of expiration.
 ///
@@ -14,15 +15,11 @@ class ExpireBuilder extends StatefulWidget {
   /// The [expireBuilder] is called when the expiration is reached, allowing you to build
   /// appropriate UI for that scenario. The [builder] is used when the app is not expired,
   /// and [loadingBuilder] is used when the expiration state is unaware.
-  ///
-  /// The [controller] is an instance of [ExpireBaseController] responsible for managing the
-  /// expiration state.
   const ExpireBuilder({
     super.key,
     required this.expireBuilder,
     required this.builder,
     required this.loadingBuilder,
-    required this.controller,
   });
 
   /// The builder function to construct the UI when the expiration is reached.
@@ -34,30 +31,30 @@ class ExpireBuilder extends StatefulWidget {
   /// The builder function to construct the UI when the expiration state is unaware.
   final Widget Function(BuildContext context) loadingBuilder;
 
-  /// The controller responsible for managing the expiration state.
-  final ExpireBaseController controller;
-
   @override
   State<ExpireBuilder> createState() => _ExpireBuilderState();
 }
 
 class _ExpireBuilderState extends State<ExpireBuilder> {
+  late ExpireBaseController _controller;
+
   @override
-  void initState() {
-    widget.controller.init();
-    super.initState();
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _controller = ExpireItProvider.of(context)!.controller;
+    _controller.init(); // Lazily call init method here
   }
 
   @override
   void dispose() {
-    widget.controller.dispose();
+    _controller.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return ValueListenableBuilder(
-      valueListenable: widget.controller.expireState,
+      valueListenable: _controller.expireState,
       builder: (context, value, child) {
         switch (value) {
           case ExpireState.expired: